+++ crates/weaver_resolver/src/dependency.rs
@@ -766,7 +766,7 @@ mod tests {
 
     fn example_v1_schema() -> V1Schema {
         V1Schema {
-            file_format: "resolved/1.0.0".to_owned(),
+            file_format: "resolved/1.0".parse().unwrap(),
             schema_url: "http://test/schemas/1.0.0".to_owned(),
             registry_id: "test-registry".to_owned(),
             registry: weaver_resolved_schema::registry::Registry {
@@ -854,7 +854,7 @@ mod tests {
 
     fn example_v2_schema() -> weaver_resolved_schema::v2::ResolvedTelemetrySchema {
         weaver_resolved_schema::v2::ResolvedTelemetrySchema {
-            file_format: "resolved/2.0".to_owned(),
+            file_format: "resolved/2.0".parse().unwrap(),
             schema_url: "http://test/schemas/2.0.0".try_into().unwrap(),
             registry: weaver_resolved_schema::v2::registry::Registry {
                 attribute_groups: vec![