+++ crates/weaver_semconv/src/v2/mod.rs
@@ -44,7 +44,11 @@ pub struct CommonFields {
     /// Specifies if the semantic convention is deprecated. The string
     /// provided as description MUST specify why it's deprecated and/or what
     /// to use instead. See also stability.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(
+        default,
+        deserialize_with = "crate::deprecated::deserialize_option_deprecated",
+        skip_serializing_if = "Option::is_none"
+    )]
     pub deprecated: Option<Deprecated>,
     /// Annotations for the attribute or signal.
     #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]